Move components to seperate files
