+++ src/components/DisplayCalc.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Display() {
+  return (
+    <div className="DisplayArea">
+      <p>0</p>
+    </div>
+  );
+}
+
+export default Display;
